move to package internal/ir
