+++ internal/ir/protocol.go
@@ -3,7 +3,7 @@
 // The full text of the Apache license is available in the LICENSE file at
 // the root of the repo.
 
-package model
+package ir
 
 type AppProtocol string
