+++ internal/xds/translator/accesslog.go
@@ -21,7 +21,6 @@ import (
 	"google.golang.org/protobuf/types/known/structpb"
 
 	"github.com/envoyproxy/gateway/internal/ir"
-	"github.com/envoyproxy/gateway/internal/model"
 	"github.com/envoyproxy/gateway/internal/utils/ptr"
 	"github.com/envoyproxy/gateway/internal/xds/types"
 )
@@ -244,7 +243,7 @@ func processClusterForAccessLog(tCtx *types.ResourceVersionTable, al *ir.AccessL
 
 		ds := &ir.DestinationSetting{
 			Weight:    ptr.To(uint32(1)),
-			Protocol:  model.GRPC,
+			Protocol:  ir.GRPC,
 			Endpoints: []*ir.DestinationEndpoint{ir.NewDestEndpoint(otel.Host, otel.Port)},
 		}
 		if err := addXdsCluster(tCtx, &xdsClusterArgs{
